made tasks module, refactored teal cron
