+++ server/src/app.module.ts
@@ -14,7 +14,7 @@ import { SimsModule } from './modules/sims/sims.module';
 import { UsersModule } from './modules/users/users.module';
 import { FileService } from './services/file.service';
 import { PrismaService } from './services/prisma.service';
-import { TealPollingService } from './services/teal-polling.service';
+import { TealPollingService } from './modules/tasks/tasks.service';
 import { SessionService } from './session/session.service';
 import { DeviceManagerModule } from './modules/device-manager/device-manager.module';
 import { InConnectModule } from './modules/in-connect/in-connect.module';
