+++ server/src/modules/tasks/tasks.module.ts
@@ -1,0 +1,10 @@
+import { Module } from '@nestjs/common';
+import { TealPollingService } from './tasks.service';
+
+
+@Module({
+    imports: [],
+    providers: [TealPollingService],
+    exports: [TealPollingService]
+})
+export class TealPollingModule { }
