+++ server/src/modules/teal-webhooks/teal-webhooks.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { TealWebhooksController } from './teal-webhooks.controller';
-import { TealPollingService } from '../../services/teal-polling.service';
+import { TealPollingService } from '../tasks/tasks.service';
 import { SimsService } from '../sims/sims.service';
 import { DataUsageService } from '../data-usage/data-usage.service';
 import { DataUsageModule } from '../data-usage/data-usage.module'; // Import the DataUsageModule
